requirements.txt
torch
transformers
accelerate
datasets
pillow
numpy
scipy
pandas
tqdm
pyyaml
